# src/app/api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mission Control API",
    "description": "AI Agent Orchestration Dashboard — REST API for managing agents, tasks, sessions, tokens, webhooks, and more.",
    "version": "1.2.0",
    "license": { "name": "MIT" },
    "contact": { "name": "Mission Control", "url": "https://github.com/builderz-labs/mission-control" }
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
  "security": [
    { "SessionCookie": [] },
    { "ApiKey": [] }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc-session",
        "description": "Session token obtained via /api/auth/login"
      },
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key (env var or DB-backed managed token)"
      }
    },
    "schemas": {
...
      "ApiToken": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "token_prefix": { "type": "string" },
          "role": { "type": "string", "enum": ["viewer", "operator", "admin"] },
          "created_by": { "type": "string" },
          "last_used_at": { "type": "integer", "nullable": true },
          "expires_at": { "type": "integer", "nullable": true },
          "revoked_at": { "type": "integer", "nullable": true },
          "created_at": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/auth/tokens": {
      "get": {
        "tags": ["Auth"],
        "summary": "List API tokens (admin only)",
        "responses": {
          "200": { "description": "List of tokens", "content": { "application/json": { "schema": { "type": "object", "properties": { "tokens": { "type": "array", "items": { "$ref": "#/components/schemas/ApiToken" } } } } } } }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Create API token (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" }, "role": { "type": "string", "enum": ["viewer", "operator", "admin"] }, "expires_in_days": { "type": "integer", "nullable": true } } } } } },
        "responses": { "201": { "description": "Token created — raw token shown only once", "content": { "application/json": { "schema": { "type": "object", "properties": { "token": { "type": "string" }, "prefix": { "type": "string" }, "id": { "type": "integer" } } } } } } }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "Rotate or revoke API token (admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["id", "action"], "properties": { "id": { "type": "integer" }, "action": { "type": "string", "enum": ["rotate", "revoke"] } } } } } },
        "responses": { "200": { "description": "Token updated" } }
      }
    },
...
    "/api/tokens": {
      "get": {
        "tags": ["Tokens"],
        "summary": "Get token usage data",
        "parameters": [
          { "name": "action", "in": "query", "schema": { "type": "string", "enum": ["list", "stats", "export", "trends", "by_agent"] } },
          { "name": "timeframe", "in": "query", "schema": { "type": "string", "enum": ["hour", "day", "week", "month", "all"] } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["json", "csv"] } }
        ],
        "responses": { "200": { "description": "Token usage data" } }
      },
      "post": {
        "tags": ["Tokens"],
        "summary": "Record token usage",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["model", "sessionId", "inputTokens", "outputTokens"], "properties": { "model": { "type": "string" }, "sessionId": { "type": "string" }, "inputTokens": { "type": "integer" }, "outputTokens": { "type": "integer" }, "operation": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Usage recorded" } }
      }
    },
...
  },
...
}